Create mapOne.txt
Txt file just to store the starting map, doesn't show where the tank is/stored as that will be handled by the color mapper. That being said, array is important for handling the bullet or handling the tank. This is because it is necessary to check what the next square holds before "shooting" or "moving", shooting a wall should destroy it. Moving into a wall should be impossible.